Fixed another compiler warning

+++ src/runtime/Classes/PBArray.m
@@ -156,7 +156,7 @@ static PBArrayValueTypeInfo PBValueTypes[] =
 - (NSString *)description
 {
 	return [NSString stringWithFormat:@"<%@ %p>{valueType = %d, count = %d, capacity = %d, data = %p}",
-			[self class], self, _valueType, _count, _capacity, _data];
+			[self class], self, _valueType, (int)_count, (int)_capacity, _data];
 }
 
 - (NSUInteger)count
